Move Helm 3 chart CRDs to the crds directory (#3206)
* Add used but missing values to chart values.yaml

imageRepositoryOverride and imageTagOverride are used but do not exist
the the values.yaml for the charts. Adding them with a default value of
an empty string preserves existing behavior but makes the values more
visible to user when referencing the values.yaml.

Signed-off-by: Michael Barker <michael.barker.84@gmail.com>

* Remove unused Helm template in _helpers.tpl

Removing the dockerRegistryOverride from _helpers.tpl for both charts as this
value is unused. The imageRepositoryOverride template appears to serve the
same purpose and is used in the charts.

Signed-off-by: Michael Barker <michael.barker.84@gmail.com>

* Move Helm 3 chart CRDs to the crds directory (#3203)

Helm charts can have dependencies on other Helm charts. When a
dependent chart has CRDs that need to be installed and the parent
chart is installing CRs using those CRDs, ordering becomes
important. In Helm 3 this ordering is handled by placing the CRDs
in a special crds directory. Helm will always install these CRDs
first.

By moving the CRDs to the crds directory, charts can include
Strimzi as a dependency without ordering issues related to
installing CRDs.

Signed-off-by: Michael Barker <michael.barker.84@gmail.com>

* Fix Javadoc header ordering

When building with JDK 14, the build fails with an error due to stricter Javadoc
rules added in JDK 13.

Signed-off-by: Michael Barker <michael.barker.84@gmail.com>

* Fix missing trailing slash in multi-line command

The script echoed by the build script when there are uncommitted changes to
generator or derived resources was missing a trailing slash causing the
command to fail when copying and pasting the command.

Signed-off-by: Michael Barker <michael.barker.84@gmail.com>

* Fix generated resources

Signed-off-by: Jakub Scholz <www@scholzj.com>

Co-authored-by: Jakub Scholz <www@scholzj.com>
Signed-off-by: Tiago Simão <tiago.simao@truphone.com>